Added user authentication feature and icons
Added Font Awesome icons and fonticons, to make the home page to look more aesthetic and better. Added icons on the login, logout, register, home and about.
I also added a user authentication when the user logs in and logs out, registers to the blog and messages asociated with the following.